rename credential file
