+++ src/controllers/validator/presentation.ts
@@ -1,5 +1,5 @@
 import { CheqdW3CVerifiablePresentation } from '../../services/w3c_presentation.js';
-import { CheqdW3CVerifiableCredentialValidator } from './cheqd-credential.js';
+import { CheqdW3CVerifiableCredentialValidator } from './credential.js';
 import { DIDValidator } from './did.js';
 import { JWTProofValidator } from './jwt-proof.js';
 import type { IValidationResult, IValidator, Validatable } from './validator.js';
